Creating a mock so that we can use it on our tests
