+++ CryptoChaser/Infrastructure/CryptoService.swift
@@ -20,27 +20,11 @@ protocol CryptoService {
     func fetchCoins() async throws -> [Currency]
 }
 
-class RemoteCryptoService: CryptoService {
+class CryptoServiceStub: CryptoService {
     var session: URLSession = URLSession(configuration: .ephemeral)
-    let fetchCoinsEnpoint = FetchCoinsEndpoint()
     
     func fetchCoins() async throws -> [Currency] {
-        guard let request = fetchCoinsEnpoint.request else {
-            throw NetworkError.badRequest
-        }
-        
-        do {
-            let (data, response) = try await session.data(for: request)
-            
-            guard let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200 else {
-                throw NetworkError.badResponse
-            }
-            
-            let decoder = JSONDecoder()
-            let coins = try decoder.decode([Currency].self, from: data)
-            return coins
-        } catch {
-            throw NetworkError.other(error)
-        }
+        let currencies = Bundle.main.decode([Currency].self, from: "Coins.json")
+        return currencies
     }
 }
